pyproject.toml
[project]
name = "gitrot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0",
    "fastapi>=0.116.1",
    "gitpython>=3.1.44",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.7",
    "langchain-openai>=0.3.27",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "uvicorn>=0.35.0",
]
